+++ treasuar/Info.plist
@@ -21,5 +21,9 @@
 			</array>
 		</dict>
 	</dict>
+	<key>UIAppFonts</key>
+	<array>
+		<string>Boldhead.otf</string>
+	</array>
 </dict>
 </plist>
