+++ treasuar/MainVC/MainVC.swift
@@ -1,0 +1,29 @@
+//
+//  MainVC.swift
+//  treasuar
+//
+//  Created by suryadi - on 28/11/22.
+//
+
+import UIKit
+
+class MainVC: UIViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+
+        // Do any additional setup after loading the view.
+    }
+
+
+    /*
+    // MARK: - Navigation
+
+    // In a storyboard-based application, you will often want to do a little preparation before navigation
+    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
+        // Get the new view controller using segue.destination.
+        // Pass the selected object to the new view controller.
+    }
+    */
+
+}
