+++ treasuar/MainVC/MainVC.xib
@@ -1,0 +1,53 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<document type="com.apple.InterfaceBuilder3.CocoaTouch.XIB" version="3.0" toolsVersion="21225" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" useSafeAreas="YES" colorMatched="YES">
+    <device id="ipad12_9" orientation="landscape" layout="fullscreen" appearance="light"/>
+    <dependencies>
+        <deployment identifier="iOS"/>
+        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="21207"/>
+        <capability name="Safe area layout guides" minToolsVersion="9.0"/>
+        <capability name="System colors in document resources" minToolsVersion="11.0"/>
+        <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
+    </dependencies>
+    <objects>
+        <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner" customClass="MainVC" customModule="treasuar" customModuleProvider="target">
+            <connections>
+                <outlet property="view" destination="i5M-Pr-FkT" id="sfx-zR-JGt"/>
+            </connections>
+        </placeholder>
+        <placeholder placeholderIdentifier="IBFirstResponder" id="-2" customClass="UIResponder"/>
+        <view clearsContextBeforeDrawing="NO" contentMode="scaleToFill" id="i5M-Pr-FkT">
+            <rect key="frame" x="0.0" y="0.0" width="1366" height="1024"/>
+            <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
+            <subviews>
+                <imageView clipsSubviews="YES" userInteractionEnabled="NO" contentMode="scaleAspectFit" horizontalHuggingPriority="251" verticalHuggingPriority="251" image="bg" translatesAutoresizingMaskIntoConstraints="NO" id="mWI-dE-Loz">
+                    <rect key="frame" x="-27" y="-24" width="1426" height="1048"/>
+                </imageView>
+                <tableView clipsSubviews="YES" contentMode="scaleToFill" alwaysBounceHorizontal="YES" showsVerticalScrollIndicator="NO" style="plain" separatorStyle="default" rowHeight="-1" estimatedRowHeight="-1" sectionHeaderHeight="-1" estimatedSectionHeaderHeight="-1" sectionFooterHeight="-1" estimatedSectionFooterHeight="-1" translatesAutoresizingMaskIntoConstraints="NO" id="baA-pj-AV1">
+                    <rect key="frame" x="20" y="276" width="1326" height="444"/>
+                    <color key="backgroundColor" systemColor="systemBackgroundColor"/>
+                    <constraints>
+                        <constraint firstAttribute="height" constant="444" id="SbD-So-dir"/>
+                        <constraint firstAttribute="width" constant="1326" id="crb-72-gNS"/>
+                    </constraints>
+                </tableView>
+            </subviews>
+            <viewLayoutGuide key="safeArea" id="fnl-2z-Ty3"/>
+            <color key="backgroundColor" systemColor="systemBackgroundColor"/>
+            <constraints>
+                <constraint firstItem="fnl-2z-Ty3" firstAttribute="trailing" secondItem="mWI-dE-Loz" secondAttribute="trailing" constant="-33" id="EAL-L5-UbO"/>
+                <constraint firstItem="mWI-dE-Loz" firstAttribute="leading" secondItem="fnl-2z-Ty3" secondAttribute="leading" constant="-27" id="EUN-cv-IB9"/>
+                <constraint firstItem="baA-pj-AV1" firstAttribute="leading" secondItem="i5M-Pr-FkT" secondAttribute="leadingMargin" id="KGO-fD-dAa"/>
+                <constraint firstItem="fnl-2z-Ty3" firstAttribute="bottom" secondItem="mWI-dE-Loz" secondAttribute="bottom" id="a4p-U5-OxG"/>
+                <constraint firstItem="baA-pj-AV1" firstAttribute="top" secondItem="fnl-2z-Ty3" secondAttribute="top" constant="256" id="bXB-iE-OZv"/>
+                <constraint firstItem="mWI-dE-Loz" firstAttribute="top" secondItem="fnl-2z-Ty3" secondAttribute="top" constant="-44" id="qWt-30-e8a"/>
+            </constraints>
+            <point key="canvasLocation" x="36.896046852122986" y="20.5078125"/>
+        </view>
+    </objects>
+    <resources>
+        <image name="bg" width="1366" height="1024"/>
+        <systemColor name="systemBackgroundColor">
+            <color white="1" alpha="1" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
+        </systemColor>
+    </resources>
+</document>
