+++ treasuar/ViewController.swift
@@ -9,6 +9,7 @@ import UIKit
 
 class ViewController: UIViewController {
     
+    @IBOutlet var ARView: UIView!
     private let imageView: UIImageView = {
         let imageView = UIImageView(frame: CGRect(x: 0, y: 0, width: 460, height: 431))
         imageView.image = UIImage(named: "SplashLogo")
@@ -16,57 +17,18 @@ class ViewController: UIViewController {
     }()
    
     
-    @IBOutlet var buttonSatu: UIButton!
-    @IBOutlet var buttonDua: UIButton!
-    @IBOutlet var buttonTiga: UIButton!
-    @IBOutlet var buttonEmpat: UIButton!
-    @IBOutlet var buttonLima: UIButton!
+
     
     override func viewDidLoad() {
         super.viewDidLoad()
         view.addSubview(imageView)
-        buttonSatu.isHidden = true
-        buttonDua.isHidden = true
-        buttonTiga.isHidden = true
-        buttonEmpat.isHidden = true
-        buttonLima.isHidden = true
+
+        
+        
         // Do any additional setup after loading the view.
     }
     
-    
-    func routeToLevelOne() {
-        guard let window = UIApplication.shared.keyWindow else { return }
-        let levelOneVC = LevelOneVC()
-        UIView.transition(with: window, duration: 0.3, options: .transitionCrossDissolve, animations: { [weak window] in
-            window?.rootViewController = levelOneVC
-        }, completion: nil)
-    }
-    
-    func routeToLevelTwo() {
-        guard let window = UIApplication.shared.keyWindow else { return }
-        let levelOneVC = LevelTwoVC()
-        UIView.transition(with: window, duration: 0.3, options: .transitionCrossDissolve, animations: { [weak window] in
-            window?.rootViewController = levelOneVC
-        }, completion: nil)
-    }
-    
-    func routeToLevelThree() {
-        guard let window = UIApplication.shared.keyWindow else { return }
-        let levelOneVC = LevelThreeVC()
-        UIView.transition(with: window, duration: 0.3, options: .transitionCrossDissolve, animations: { [weak window] in
-            window?.rootViewController = levelOneVC
-        }, completion: nil)
-    }
-    
-    @IBAction func oneAction(_ sender: Any) {
-        routeToLevelOne()
-    }
-    @IBAction func twoAction(_ sender: Any) {
-        routeToLevelTwo()
-    }
-    @IBAction func threeAction(_ sender: Any) {
-        routeToLevelThree()
-    }
+
     
     override func viewDidLayoutSubviews() {
         super.viewDidLayoutSubviews()
@@ -91,11 +53,7 @@ class ViewController: UIViewController {
         } completion: { isComplete in
             if isComplete{
                 self.imageView.alpha = 0
-                self.buttonSatu.isHidden = false
-                self.buttonDua.isHidden = false
-                self.buttonTiga.isHidden = false
-                self.buttonEmpat.isHidden = false
-                self.buttonLima.isHidden = false
+
             }
             
         }
